bbu_reference/bbu2-master/pyproject.toml
[project]
name = "bbu2"
version = "0.1.0"
description = "Add your description here"
authors = [
    {name = "erzhan", email = "erzhan.sarsenov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "pydantic-settings>=2.6.1",
    "requests>=2.32.3",
    "pytelegrambotapi>=4.24.0",
    "pydantic>=2.10.1",
    "pybit>=5.8.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
